Fix final lint errors

+++ ros1/src/vision/march_foot_position_finder/src/point_finder.cpp
@@ -21,6 +21,7 @@ using PointCloud = pcl::PointCloud<Point>;
  * foot
  * @param step_point an initial desired step point
  */
+// Suppress lint error "variables are not initialized" (ros parameters)
 // NOLINTNEXTLINE
 PointFinder::PointFinder(ros::NodeHandle* n, PointCloud::Ptr pointcloud,
     std::string left_or_right, Point& step_point)
@@ -183,14 +184,15 @@ void PointFinder::findFeasibleFootPlacements(std::vector<Point>* position_queue)
                 }
             }
 
-            // Suppress lint errors: narrowing conversions (they are safe)
+            // Suppress lint errors "narrowing conversions" (they are safe)
             if (num_free_cells
                 >= rect_height_ * rect_width_ * available_points_ratio_) {
-                float x = ((float)x_opt / (float)grid_resolution_)
-                    - (float)x_offset_ + (float)cell_width_ / 2.0;
-                float y = (((float)grid_resolution_ - (float)y_opt)
-                              / (float)grid_resolution_)
-                    - (float)y_offset_ - (float)cell_width_ / 2.0;
+                // NOLINTNEXTLINE
+                float x = ((float)x_opt / grid_resolution_) - x_offset_
+                    + cell_width_ / 2.0;
+                // NOLINTNEXTLINE
+                float y = ((float)(grid_resolution_ - y_opt) / grid_resolution_)
+                    - y_offset_ - cell_width_ / 2.0;
                 float z = height_map_[y_opt][x_opt];
 
                 if (std::abs(z - current_foot_z_) <= max_z_distance_
